BACKEND/requirements.txt
fastapi
uvicorn
pymongo
motor
python-dotenv
google-generativeai
pydantic